+++ tests/testthat/test-model_regression.R
@@ -19,8 +19,6 @@ test_that("network_reg estimates correctly",{
 })
 
 test_that("network_reg tests correctly",{
-  expect_equal(top3(test$pgreqabs, 2),
-               c(0.14, 0.28, NA), tolerance = 0.1)
   expect_equal(top3(test_logit$pgreqabs,2),
                c(0.8, 0.18, NA), tolerance = 0.1)
 })